+++ src/main/java/de/jeff_media/AngelChestPlus/config/ConfigUtils.java
@@ -158,6 +158,7 @@ public class ConfigUtils {
 		conf.addDefault(Config.SHOW_MESSAGE_WHEN_OTHER_PLAYER_EMPTIES_CHEST, true);
 		conf.addDefault(Config.SHOW_MESSAGE_WHEN_OTHER_PLAYER_OPENS_CHEST, true);
 		conf.addDefault(Config.PREFIX_MESSAGES, true);
+		conf.addDefault(Config.NEVER_REPLACE_BEDROCK,true);
 		main.disabledMaterials = conf.getStringList(Config.DISABLED_MATERIALS);
 		main.disabledWorlds =  conf.getStringList(Config.DISABLED_WORLDS);
 		main.disabledRegions =  conf.getStringList(Config.DISABLED_WORLDGUARD_REGIONS);